file structure to add navigation to storybook
