+++ app/frontend/src/components/Navigation/index.ts
@@ -1,0 +1,2 @@
+export * from "./Navigation";
+export { default } from "./Navigation";
