+++ app/frontend/storybook-static/asset-manifest.json
@@ -1,0 +1,29 @@
+{
+  "files": {
+    "main.css": "./static/css/main.ced0f865.chunk.css",
+    "main.js": "./main.62a63e520da5d4ae014e.bundle.js",
+    "main.css.map": "./static/css/main.ced0f865.chunk.css.map",
+    "main.js.map": "./main.62a63e520da5d4ae014e.bundle.js.map",
+    "runtime~main.js": "./runtime~main.62a63e520da5d4ae014e.bundle.js",
+    "runtime~main.js.map": "./runtime~main.62a63e520da5d4ae014e.bundle.js.map",
+    "vendors~main.js": "./vendors~main.62a63e520da5d4ae014e.bundle.js",
+    "vendors~main.js.map": "./vendors~main.62a63e520da5d4ae014e.bundle.js.map",
+    "iframe.html": "./iframe.html",
+    "static/media/code-brackets.2e1112d7.svg": "./static/media/code-brackets.2e1112d7.svg",
+    "static/media/colors.a4bd0486.svg": "./static/media/colors.a4bd0486.svg",
+    "static/media/comments.a3859089.svg": "./static/media/comments.a3859089.svg",
+    "static/media/direction.b770f9af.svg": "./static/media/direction.b770f9af.svg",
+    "static/media/flow.edad2ac1.svg": "./static/media/flow.edad2ac1.svg",
+    "static/media/funnycat.a51bf746.jpg": "./static/media/funnycat.a51bf746.jpg",
+    "static/media/plugin.d494b228.svg": "./static/media/plugin.d494b228.svg",
+    "static/media/repo.6d496322.svg": "./static/media/repo.6d496322.svg",
+    "static/media/stackalt.dba9fbb3.svg": "./static/media/stackalt.dba9fbb3.svg",
+    "vendors~main.62a63e520da5d4ae014e.bundle.js.LICENSE.txt": "./vendors~main.62a63e520da5d4ae014e.bundle.js.LICENSE.txt"
+  },
+  "entrypoints": [
+    "runtime~main.62a63e520da5d4ae014e.bundle.js",
+    "vendors~main.62a63e520da5d4ae014e.bundle.js",
+    "static/css/main.ced0f865.chunk.css",
+    "main.62a63e520da5d4ae014e.bundle.js"
+  ]
+}
